src/EducativeIo/GrokkingTheCodingInterview/Ch02_Pattern_SlidingWindow/P7_LongestSubarrayWithOnesAfterReplacement/Java/Solution.java
package EducativeIo.GrokkingTheCodingInterview.Ch02_Pattern_SlidingWindow.P7_LongestSubarrayWithOnesAfterReplacement.Java;

public class Solution {
    public static void main(String[] args) {
        System.out.println(findLength(new int[] { 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1 }, 2));
        System.out.println(findLength(new int[] { 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1 }, 3));
    }

    public static int findLength(int[] arr, int k) {
        int maxLength = Integer.MIN_VALUE;
        int windowStart = 0;
        int maxOnesCount = 0;
        for (int windowEnd = 0; windowEnd < arr.length; windowEnd++) {
            if (arr[windowEnd] == 1) {
                maxOnesCount++;
            }

            // current window size is from windowStart to windowEnd, overall we have a maximum of 1s
            // repeating a maximum of 'maxOnesCount' times, this means that we can have a window with
            // 'maxOnesCount' 1s and the remaining are 0s which should replace with 1s.
            // now, if the remaining 0s are more than 'k', it is the time to shrink the window as we
            // are not allowed to replace more than 'k' Os
            if (windowEnd - windowStart + 1 - maxOnesCount > k) {
                if (arr[windowStart] == 1) {
                    maxOnesCount--;
                }
                windowStart++;
            }
            maxLength = Math.max(maxLength, windowEnd - windowStart + 1);
        }
        return maxLength;
    }
}